POC: rename network_layer.py to network_layer_shared.py
Rename pymongo/network_layer.py to pymongo/network_layer_shared.py to make
explicit that this file is intentionally shared between async and sync code,
and is not processed by synchro. Update all 9 import sites.

Add clarifying module docstrings to periodic_executor.py and lock.py
explaining why they also live in the top-level package alongside both
async and sync implementations.
