+++ pymongo/asynchronous/encryption.py
@@ -81,7 +81,7 @@ from pymongo.errors import (
     ServerSelectionTimeoutError,
 )
 from pymongo.helpers_shared import _get_timeout_details
-from pymongo.network_layer import async_socket_sendall
+from pymongo.network_layer_shared import async_socket_sendall
 from pymongo.operations import UpdateOne
 from pymongo.pool_options import PoolOptions
 from pymongo.pool_shared import (
@@ -211,7 +211,7 @@ class _EncryptionIO(AsyncMongoCryptCallback):  # type: ignore[misc]
                     if _IS_SYNC:
                         data = conn.recv(kms_context.bytes_needed)
                     else:
-                        from pymongo.network_layer import (  # type: ignore[attr-defined]
+                        from pymongo.network_layer_shared import (  # type: ignore[attr-defined]
                             async_receive_data_socket,
                         )
